fix(app-vite): Cordova & Android -> usesCleartextTraffic attribute is always true in the AndroidManifest.xml file #17219
